--- src/sidebar/templates/thread-list.html
@@ -1,18 +1,0 @@
-<ul class="thread-list">
-  <li class="thread-list__spacer"
-      ng-style="{height: vm.virtualThreadList.offscreenUpperHeight}"></li>
-  <li ng-repeat="child in vm.virtualThreadList.visibleThreads track by child.id">
-    <div id="{{child.id}}"
-        class="thread-list__card"
-        ng-mouseenter="vm.onFocus({annotation: child.annotation})"
-        ng-class="{'thread-list__card--theme-clean' : vm.isThemeClean }"
-        ng-click="vm.onSelect({annotation: child.annotation})"
-        ng-mouseleave="vm.onFocus({annotation: null})">
-        <thread thread="child" show-document-info="vm.showDocumentInfo"></thread>
-    </div>
-    <hr ng-if="vm.isThemeClean"
-        class="thread-list__separator--theme-clean" />
-  </li>
-  <li class="thread-list__spacer"
-      ng-style="{height: vm.virtualThreadList.offscreenLowerHeight}"></li>
-</ul>
