--- src/styles/sidebar/components/thread-list.scss
@@ -1,41 +1,0 @@
-@use "../../variables" as var;
-
-.thread-list {
-  & > * {
-    // Default spacing between items in the annotation card list
-    margin-bottom: 0.72em;
-  }
-}
-
-.thread-list__card {
-  box-shadow: 0px 1px 1px 0px rgba(0, 0, 0, 0.1);
-  border-radius: 2px;
-  cursor: pointer;
-  padding: 1em;
-  background-color: var.$white;
-
-  &:hover {
-    box-shadow: 0px 2px 3px 0px rgba(0, 0, 0, 0.15);
-  }
-}
-
-.thread-list__card--theme-clean {
-  box-shadow: none;
-
-  &:hover {
-    box-shadow: none;
-  }
-}
-
-.thread-list__spacer {
-  // This is a hidden element which is used to reserve space for off-screen
-  // threads, so it should not occupy any space other than that set via its
-  // 'height' inline style property.
-  margin: 0;
-}
-
-.thread-list__separator--theme-clean {
-  border: 0;
-  border-top: 1px solid #e1e1e1;
-  margin: -8px 15px 10px 15px;
-}
