+++ src/styles/sidebar/sidebar.scss
@@ -63,7 +63,6 @@
 @use './components/tag-list';
 @use './components/thread';
 @use './components/thread-card';
-@use './components/thread-list';
 @use './components/toast-messages';
 @use './components/tooltip';
 @use './components/top-bar';
